% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_series}
\alias{plot_series}
\title{Plot the two series on a common time axis}
\usage{
plot_series(x, unit = "period")
}
\arguments{
\item{x}{A \code{lag_ceiling} object.}

\item{unit}{Label for the time axis.}
}
\value{
A ggplot.
}
\description{
Reproduction (blue) and recruitment (vermillion) as bars, one panel each,
with the concentration of each printed in the panel title.
}
\examples{
ce <- lag_ceiling(lepanthes_census, reproduction = "inflorescences", K = 4,
                  kernels = lag_kernels(4, bin = 6, unit = "mo"))
plot_series(ce, unit = "census")
}
